requirements.txt
nats_py==2.3.1
PyYAML==6.0.1
PyYAML==6.0.1
